+++ lib/project_types/node/commands/serve.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 module Node
   class Command
-    class Serve < ShopifyCLI::SubCommand
+    class Serve < ShopifyCLI::Command::AppSubCommand
       prerequisite_task ensure_project_type: :node
       prerequisite_task :ensure_env, :ensure_dev_store