+++ lib/shopify_cli/command/app_sub_command.rb
@@ -1,0 +1,11 @@
+module ShopifyCLI
+  class Command
+    class AppSubCommand < SubCommand
+      class << self
+        def call_help(*)
+          @ctx.puts(help)
+        end
+      end
+    end
+  end
+end