+++ test/project_types/node/commands/tunnel_test.rb
@@ -11,22 +11,22 @@ module Node
 
       def test_auth
         ShopifyCLI::Tunnel.any_instance.expects(:auth)
-        run_cmd("node tunnel auth adfhauf98q7rtqhfkajf")
+        run_cmd("app node tunnel auth adfhauf98q7rtqhfkajf")
       end
 
       def test_auth_no_token
         ShopifyCLI::Tunnel.any_instance.expects(:auth).never
-        run_cmd("node tunnel auth")
+        run_cmd("app node tunnel auth")
       end
 
       def test_start
         ShopifyCLI::Tunnel.any_instance.expects(:start)
-        run_cmd("node tunnel start")
+        run_cmd("app node tunnel start")
       end
 
       def test_stop
         ShopifyCLI::Tunnel.any_instance.expects(:stop)
-        run_cmd("node tunnel stop")
+        run_cmd("app node tunnel stop")
       end
     end
   end