+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 Unreleased
 ------
 * [#1566](https://github.com/Shopify/shopify-cli/pull/1566): Fix bug when running `npm | yarn list` for extension package resolution
-
+* [#1557](https://github.com/Shopify/shopify-cli/pull/1557): **Breaking** Move app commands under `shopify app`.
+  
 Version 2.5.0
 ------